feat: update dashboard components for improved functionality and structure
